3.0FastAPI/requirements.txt
fastapi
uvicorn
motor
beanie
pydantic
python-dotenv

=>
pip install -r requirements.txt